// api/models/ExpenseSummaryDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ExpenseSummaryDto = {
    total?: number;
    food?: number;
    transport?: number;
    entertainment?: number;
    housing?: number;
    health?: number;
    shopping?: number;
    education?: number;
    travel?: number;
    utilities?: number;
    other?: number;
};