// api/models/TransactionCategory.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum TransactionCategory {
    FOOD = 'Food',
    TRANSPORT = 'Transport',
    ENTERTAINMENT = 'Entertainment',
    HOUSING = 'Housing',
    HEALTH = 'Health',
    SHOPPING = 'Shopping',
    EDUCATION = 'Education',
    TRAVEL = 'Travel',
    UTILITIES = 'Utilities',
    OTHER = 'Other',
}