// api/models/YearlyTransactionsResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { MonthlyTransactionSummaryDto } from './MonthlyTransactionSummaryDto';
export type YearlyTransactionsResponseDto = {
    year?: number;
    transactions?: Array<MonthlyTransactionSummaryDto> | null;
};